+++ test/browser/features/fixtures/packages/angular/src/app/app.module.ts
@@ -1,0 +1,41 @@
+import { HttpClientModule } from '@angular/common/http';
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
+
+import { AppRoutingModule } from './app-routing.module';
+import { AppComponent } from './app.component';
+import BugsnagPerformance from '@bugsnag/browser-performance';
+import { AngularRoutingProvider, bugsnagBootstrapper } from '@bugsnag/angular-performance';
+import { ProfileComponent } from './profile/profile.component';
+
+const parameters = new URLSearchParams(window.location.search)
+const apiKey = parameters.get('api_key')!
+const endpoint = parameters.get('endpoint')!
+
+BugsnagPerformance.start({
+  apiKey,
+  endpoint,
+  // @ts-expect-error undocumented config for testing purposes
+  maximumBatchSize: 16,
+  batchInactivityTimeoutMs: 5000,
+  autoInstrumentNetworkRequests: false,
+  autoInstrumentRouteChanges: true,
+  routingProvider: new AngularRoutingProvider(),
+})
+
+@NgModule({
+  declarations: [
+    AppComponent,
+    ProfileComponent
+  ],
+  imports: [
+    BrowserModule,
+    HttpClientModule,
+    AppRoutingModule
+  ],
+  providers: [
+    bugsnagBootstrapper,
+  ],
+  bootstrap: [AppComponent]
+})
+export class AppModule { }